requirements.txt

ipykernel==7.2.0
pandas==2.3.3
python-dotenv==1.2.1
requests==2.32.5
sphinx-ai-cli==1.0.6
streamlit==1.54.0


